filename changes, more python 2 to 3 changes
