+++ heteropolymer_fit/generate_fitting_eqns_heteropolymer.py
