+++ heteropolymer_fit/ising_fitter_heteropolymer.py
