+++ homopolymer_fit/data_conversion_aviv.py
@@ -1,22 +1,18 @@
 """
 Converts Aviv.dat files into numpy files for NRC capped homopolymer repeats.
-
-Most recent revision: 05/21/2020, Doug Barrick
-
 """
 
 import numpy as np
 import pandas as pd
 import ntpath  # Good for path manipulations on a PC?
 import glob  # Allows for unix-like specifications paths, using *, ?, etc.
-import csv
+import os
 import json
 import time
 
 start = time.time()
 
-path = "/Users/dougbarrick/OneDrive - Johns Hopkins/Manuscripts/Ising_program/\
-Scripts/scripts_vanilla_NRC/cANK_scripts/cANK_spyder_scripts/"
+PATH = os.path.dirname(os.path.abspath(__file__))
 proj_name = "cANK"
 
 den_nsig_const_melt = []
@@ -33,7 +29,7 @@ den_nsig_const_melt_df = pd.DataFrame(
 
 # Gets file names, and extracts information including construct name, melt number.
 num = 0
-for filename in glob.glob("{}*.dat".format(path)):
+for filename in glob.glob("{}*.dat".format(PATH)):
     num = num + 1
     base = ntpath.basename(filename)
     melt = base.split(".")[0]
@@ -82,7 +78,7 @@ for filename in glob.glob("{}*.dat".format(path)):
             )
         melt_array = np.array(single_melt_dncm)
         np.save(
-            path + melt, melt_array
+            PATH + melt, melt_array
         )  # Writes an npy file to disk for each melt.
         temp_df = pd.DataFrame(melt_array)
         den_nsig_const_melt_df = den_nsig_const_melt_df.append(temp_df)
@@ -91,7 +87,7 @@ for filename in glob.glob("{}*.dat".format(path)):
         melts.append(melt)
 
 den_nsig_const_melt_df.to_csv(
-    "{}{}_combined_data.csv".format(path, proj_name), index=False, header=False
+    "{}{}_combined_data.csv".format(PATH, proj_name), index=False, header=False
 )
 
 # This loop puts melts in order of type (NRxC, NRx, RxC) and length.
@@ -117,10 +113,10 @@ for melt in melts:
         constructs.append(construct)
 
 # Write out the results.
-with open("{0}{1}_constructs.txt".format(path, proj_name), "wb") as r:
+with open("{0}{1}_constructs.txt".format(PATH, proj_name), "wb") as r:
     json.dump(constructs, r)
 
-with open("{0}{1}_melts.txt".format(path, proj_name), "wb") as s:
+with open("{0}{1}_melts.txt".format(PATH, proj_name), "wb") as s:
     json.dump(melts, s)
 
 stop = time.time()
